Refatorado código e criado pacote de teste.
